+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -47,9 +47,9 @@ const cli = {
         describe: 'A flag to disable the selenium docker service.',
         default: false,
       },
-      disableSnapshotMismatch: {
+      ignoreScreenshotMistmatch: {
         type: 'boolean',
-        describe: 'A flag to disable failing on snapshot mismatch.',
+        describe: 'A flag to disable failing on screenshot mismatch.',
         default: false,
       },
       externalHost: {