+++ packages/terra-functional-testing/tests/wdio/terra-validates-spec.js
@@ -37,13 +37,13 @@ Terra.describeViewports('Terra.validates', ['small', 'large'], () => {
       Terra.validates.screenshot('validates-screenshot');
     });
 
-    it('should pass if disableSnapshotMismatch is true even when mismatched', () => {
-      const oldDisableSnapshotMismatch = Terra.serviceOptions.disableSnapshotMismatch;
-      Terra.serviceOptions.disableSnapshotMismatch = true;
+    it('should pass if ignoreScreenshotMistmatch is true even when mismatched', () => {
+      const oldIgnoreScreenshotMistmatch = Terra.serviceOptions.ignoreScreenshotMistmatch;
+      Terra.serviceOptions.ignoreScreenshotMistmatch = true;
       browser.url('/insufficient-color-contrast.html');
 
       Terra.validates.screenshot('validates-mismatch');
-      Terra.serviceOptions.disableSnapshotMismatch = oldDisableSnapshotMismatch;
+      Terra.serviceOptions.ignoreScreenshotMistmatch = oldIgnoreScreenshotMistmatch;
     });
 
     it('should require a screenshot name', () => {